Move decode_transcript to data_structures.py
